Refs #136766 added TemporalWidgetView used as widget for viewing temporal values
