+++ src/components/index.js
@@ -3,3 +3,7 @@
  * @module components
  */
 export TemporalWidget from './manage/Widgets/TemporalWidget';
+
+export TemporalWidgetEdit from './manage/Widgets/TemporalWidget';
+
+export TemporalWidgetView from './theme/Widgets/TemporalWidget';
