+++ src/components/theme/Widgets/public.less
@@ -1,0 +1,7 @@
+.temporal-coverage.widget {
+  display: flex;
+  flex-flow: wrap;
+}
+.temporal-coverage-value {
+  text-indent: 15px;
+}
