+++ src/index.js
@@ -1,9 +1,9 @@
-import { TemporalWidget } from './components';
+import { TemporalWidgetEdit } from './components';
 
 const applyConfig = (config) => {
   config.widgets.widget = {
     ...config.widgets.widget,
-    temporal: TemporalWidget,
+    temporal: TemporalWidgetEdit,
   };
   return config;
 };
